Implement paging

+++ app/src/main/java/com/monsjoker/namadaexplorer/uis/screens/home/data/HomeDetailsData.kt
@@ -2,7 +2,7 @@ package com.monsjoker.namadaexplorer.uis.screens.home.data
 
 data class HomeDetailsData(
     val epoch: Int = 0,
-    val blockHeight: Int,
+    val blockHeight: Long,
     val totalStake: Int,
     val validators: Int,
     val governanceProposals: Int = 0
